--- a/.gitignore
+++ b/.gitignore
@@ -5,3 +5,4 @@
 Manifest.toml
 *.DS_Store
 .vscode
+profile.pb.gz