--- a/src/bdf_utils.jl
+++ b/src/bdf_utils.jl
@@ -147,15 +147,18 @@ end
 
 function calc_Lagrange_interp!(k,weights,t,ts,u_history,u)
   vc = _vec(u)
-  if t in ts
-    i = searchsortedfirst(ts,t,rev=true)
+  # ts is short enough that linear search will be faster
+  i = findfirst(isequal(t), ts)
+  if i !== nothing
     @.. @views vc = u_history[:,i]
   else
     for i in 1:k+1
-      @.. @views vc += weights[i]/(t-ts[i])*u_history[:,i]
+      wdt = weights[i]/(t-ts[i])
+      @.. @views vc = muladd(wdt, u_history[:,i], vc)
     end
     for i in 1:k+1
-      @.. u *= t-ts[i]
+      dt = t-ts[i]
+      @.. u *= dt
     end
   end
 end
@@ -206,29 +209,30 @@ function reinitFBDF!(integrator, cache)
     cache.iters_from_event = 0
   end
 
-  if iters_from_event == 0
+  vuprev = _vec(uprev)
+  @views if iters_from_event == 0
     weights[1] = 1/dt
     ts[1] = t
-    @.. u_history[:,1] = $_vec(uprev)
+    @.. u_history[:,1] = vuprev
   elseif iters_from_event == 1
     ts[2] = ts[1]
     ts[1] = t
-    @.. @views u_history[:,2] = u_history[:,1]
-    @.. u_history[:,1] = $_vec(uprev)
+    @.. u_history[:,2] = u_history[:,1]
+    @.. u_history[:,1] = vuprev
   elseif consfailcnt == 0
     for i in order+2:-1:2
       ts[i] = ts[i-1]
-      @.. @views u_history[:,i] = u_history[:,i-1]
+      @.. u_history[:,i] = u_history[:,i-1]
     end
     ts[1] = t
-    @.. u_history[:,1] = $_vec(uprev)
+    @.. u_history[:,1] = vuprev
   end
   if iters_from_event >= 1
     compute_weights!(ts,order,weights)
   end
 end
 
-function estimate_terk!(integrator, cache, k, max_order)
+function estimate_terk!(integrator, cache, k, ::Val{max_order}) where max_order
   #calculate hᵏ⁻¹yᵏ⁻¹
   @unpack ts_tmp, terk_tmp, u_history = cache
   @unpack t, dt, u = integrator
@@ -241,7 +245,7 @@ function estimate_terk!(integrator, cache, k, max_order)
   @.. terk_tmp *= abs(dt^(k-1))
 end
 
-function estimate_terk(integrator, cache, k, max_order, u)
+function estimate_terk(integrator, cache, k, ::Val{max_order}, u) where max_order
   @unpack ts_tmp, u_history = cache
   @unpack t, dt = integrator
   fd_weights = calc_finite_difference_weights(ts_tmp,t+dt,k-1,Val(max_order))