--- a/test/interface/mass_matrix_tests.jl
+++ b/test/interface/mass_matrix_tests.jl
@@ -80,6 +80,7 @@ dependent_M2 = DiffEqArrayOperator(ones(3,3),update_func=update_func2)
     @test _norm_dsol(QBDF1(),prob,prob2) ≈ 0 atol=1e-12
     @test _norm_dsol(QBDF2(),prob,prob2) ≈ 0 atol=1e-12
     @test _norm_dsol(QBDF(),prob,prob2) ≈ 0 atol=1e-12
+    @test _norm_dsol(FBDF(),prob,prob2) ≈ 0 atol=1e-12
 
     @test _norm_dsol(QNDF1(),prob,prob2) ≈ 0 atol=1e-12
     @test _norm_dsol(QNDF2(),prob,prob2) ≈ 0 atol=1e-12