Master:

```julia
julia> @benchmark solve($prob,FBDF())
BenchmarkTools.Trial: 8173 samples with 1 evaluation.
 Range (min … max):  495.030 μs …   6.687 ms  ┊ GC (min … max): 0.00% … 89.02%
 Time  (median):     551.189 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   608.267 μs ± 308.665 μs  ┊ GC (mean ± σ):  3.03% ±  5.52%

       ██▆▄▃▃▃▄▃▃▂▂▁▁▁▁▁▁▁ ▁ ▁▁                                 ▂
  ▇▇▅▅███████████████████████████████▇▇███▆▆▆▇▆▆▇▅▆▆▅▆▆▆▅▅▃▅▄▄▄ █
  495 μs        Histogram: log(frequency) by time        993 μs <

 Memory estimate: 252.56 KiB, allocs estimate: 3145.
```

PR:
```julia
julia> @benchmark solve($prob,FBDF())
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  196.784 μs …   5.117 ms  ┊ GC (min … max): 0.00% … 94.63%
 Time  (median):     204.281 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   217.876 μs ± 159.520 μs  ┊ GC (mean ± σ):  2.37% ±  3.13%

  ▄██▇▇▅▄▃▂▁▂▃▃▃▃▃▃▃▂▂▁▁ ▁▁                                     ▂
  █████████████████████████████▇▇▇▆▆▆▆▇▆▆▅▆▆▆▆▃▅▅▅▅▄▄▃▄▄▄▄▃▅▄▄▄ █
  197 μs        Histogram: log(frequency) by time        328 μs <

 Memory estimate: 94.69 KiB, allocs estimate: 542.
```